+++ src/constants/dump-event.ts
@@ -1,1 +1,2 @@
-export const DUMP_EVENT: string = 'dump';
+export const WRITE_EVENT: string = 'write';
+export const READ_EVENT: string = 'read';